Trust Tokens: Add an XMLHttpRequest interface
The experimental Trust Token API
(https://github.com/wicg/trust-token-api)
takes in parameters specifying a Trust Tokens protocol operation,
annotates outgoing requests with request headers corresponding to
protocol state, and processes corresponding response headers. For the
MVP of the API, we're making the functionality available via Fetch,
iframe, and XHR.

To implement the XHR surface, this CL adds a new XHR setter,
that accepts a trustToken object (identical to the one for the Fetch
interface added in crrev.com/c/2036648) specifying a Trust Tokens
operation to execute against the request.

Test: Expands integration tests to ensure Trust Tokens params are
propagated correctly from the XHR interface to the network stack.
Bug: 1062395
Change-Id: I5f32d60c7b59002e79dc877719ba549ce5b9f412