Added Badging API Web Platform Tests.
The [Badging API](https://wicg.github.io/badging/) is a new API
allowing the setting of application/document badges. This change
converts the Badging API web tests into Web Platform Tests.

This change adds a no-op MockBadgeService to the content_shell.
This is necessary because the RenderProcessHostImpl will terminate
the render process if unknown Mojo messages are received.

Bug: 1051684
Change-Id: I61e29b8f197a754fea489ebb2d0dcd41484f6cf4